+++ data-transfer/src/main/java/it/eng/datatransfer/service/DataTransferService.java
@@ -24,6 +24,7 @@ import it.eng.datatransfer.repository.TransferRequestMessageRepository;
 import it.eng.datatransfer.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
 import it.eng.tools.controller.ApiEndpoints;
+import it.eng.tools.model.IConstants;
 import it.eng.tools.response.GenericApiResponse;
 import lombok.extern.slf4j.Slf4j;
 
@@ -96,7 +97,7 @@ public class DataTransferService {
 				.format(transferRequestMessage.getFormat())
 				.dataAddress(transferRequestMessage.getDataAddress())
 				.state(TransferState.REQUESTED)
-				.role(transferProcess.getRole())
+				.role(IConstants.ROLE_PROVIDER)
 				.datasetId(transferProcess.getDatasetId())
 				.createdBy(transferProcess.getCreatedBy())
 				.lastModifiedBy(transferProcess.getLastModifiedBy())
@@ -158,8 +159,16 @@ public class DataTransferService {
 		log.debug("Starting data transfer for consumerPid {} and providerPid {}", consumerPidFinal, providerPidFinal);
 
 		TransferProcess transferProcessRequested = findTransferProcess(consumerPidFinal, providerPidFinal);
+		
+		if (IConstants.ROLE_PROVIDER.equals(transferProcessRequested.getRole()) && TransferState.REQUESTED.equals(transferProcessRequested.getState())) {
+			// Only consumer can transit from REQUESTED to STARTED state
+			throw new TransferProcessInvalidStateException("State transition aborted, consumer can not transit from " + TransferState.REQUESTED.name()
+					+ " to " + TransferState.STARTED.name(),
+					transferProcessRequested.getConsumerPid(), transferProcessRequested.getProviderPid());
+		} 
+		
 		stateTransitionCheck(transferProcessRequested, TransferState.STARTED);
-
+		
 		TransferProcess transferProcessStarted = transferProcessRequested.copyWithNewTransferState(TransferState.STARTED);
 		transferProcessRepository.save(transferProcessStarted);
 		publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
@@ -254,7 +263,7 @@ public class DataTransferService {
 	public TransferProcess findTransferProcess(String consumerPid, String providerPid) {
 		TransferProcess transferProcessRequested = transferProcessRepository.findByConsumerPidAndProviderPid(consumerPid, providerPid)
 			.orElseThrow(() -> new TransferProcessNotFoundException("Transfer process for consumerPid " + consumerPid
-			 + " and providerPid " + consumerPid + " not found"));
+			 + " and providerPid " + providerPid + " not found"));
 		return transferProcessRequested;
 	}